+++ src/Lab1/Lab1/Commands/AddFunction.cs
@@ -1,0 +1,22 @@
+﻿using Spectre.Console.Cli;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+
+
+namespace Lab1.Commands
+{
+    class AddFunction : Command<AddFigureCommand.AddFigureSettings>
+    {
+        public class AddFigureSettings : CommandSettings
+        {
+
+        }
+
+
+
+    }
+}
